+++ common/lib/llm/clients/lmstudio_client.py
@@ -10,16 +10,15 @@ This class is primarily intended for interfacing with LiteLLM, but since
 LiteLLM itself is mostly OpenAI API-compatible, this can be used to interface
 with the OpenAI API as well.
 """
-import requests
-
 from common.lib.llm.llm_client import LLMProviderClient
 
-class LiteLLMClient(LLMProviderClient):
-    type = "litellm"
 
-    _models_info_path = "/model/info"
-    _models_info_key = "data"
-    _model_id_key = "model_name"
+class LMStudioClient(LLMProviderClient):
+    type = "lmstudio"
+
+    _models_info_path = "/api/v1/models"
+    _models_info_key = "models"
+    _model_id_key = "key"
 
     def parse_supported_media_types(self, meta: dict) -> list[str]:
         """
@@ -29,11 +28,11 @@ class LiteLLMClient(LLMProviderClient):
         :returns:       Ordered list of supported media type strings.
                         Returns ``[]`` when ``meta`` is ``None``
         """
-        if meta is None or not meta.get("model_info"):
+        if meta is None or not meta.get("capabilities"):
             return []
 
         media_types = {"text"}  # far as I can tell, text is always supported
-        if meta["model_info"].get("supports_vision"):
+        if meta["capabilities"].get("vision"):
             media_types.add("image")
 
         if meta["model_info"].get("supports_audio_input"):
@@ -53,10 +52,16 @@ class LiteLLMClient(LLMProviderClient):
         """
         model_name = self.get_global_model_id(meta)
 
-        if meta.get("model_name"):
-            model_name = meta["model_name"]
+        if meta.get("display_name"):
+            model_name = meta["display_name"]
 
-        if meta["litellm_params"].get("model"):
-            model_name = "/".join(meta["litellm_params"].get("model").split("/")[1:])
+        extra_bits = []
+        if "publisher" in meta:
+            extra_bits.append(meta["publisher"])
+
+        if "params_string" in meta:
+            extra_bits.append(meta["params_string"])
+
+        model_name += f" {', '.join(extra_bits)}"
 
         return model_name
