+++ common/lib/llm/clients/vllm_client.py
@@ -1,25 +1,21 @@
 """
-Centralized HTTP client for communicating with a LiteLLM server.
+Centralized HTTP client for communicating with a vLLM server.
 
-This class owns all direct HTTP calls to LiteLLM's REST API and provides shared
+This class owns all direct HTTP calls to vLLM's REST API and provides shared
 static helpers for capability parsing, display-name formatting, and building
 canonical llm.available_models entries. It is a plain helper with no 4CAT
 base-class dependency.
 
-This class is primarily intended for interfacing with LiteLLM, but since
-LiteLLM itself is mostly OpenAI API-compatible, this can be used to interface
-with the OpenAI API as well.
 """
-import requests
-
 from common.lib.llm.llm_client import LLMProviderClient
 
-class LMStudioClient(LLMProviderClient):
-    type = "lmstudio"
 
-    _models_info_path = "/api/v1/models"
-    _models_info_key = "models"
-    _model_id_key = "key"
+class VLLMClient(LLMProviderClient):
+    type = "vllm"
+
+    _models_info_path = "/model/info"
+    _models_info_key = "data"
+    _model_id_key = "model"
 
     def parse_supported_media_types(self, meta: dict) -> list[str]:
         """
@@ -29,11 +25,11 @@ class LMStudioClient(LLMProviderClient):
         :returns:       Ordered list of supported media type strings.
                         Returns ``[]`` when ``meta`` is ``None``
         """
-        if meta is None or not meta.get("capabilities"):
+        if meta is None or not meta.get("model_info"):
             return []
 
         media_types = {"text"}  # far as I can tell, text is always supported
-        if meta["capabilities"].get("vision"):
+        if meta["model_info"].get("supports_vision"):
             media_types.add("image")
 
         if meta["model_info"].get("supports_audio_input"):
@@ -53,17 +49,10 @@ class LMStudioClient(LLMProviderClient):
         """
         model_name = self.get_global_model_id(meta)
 
-        if meta.get("display_name"):
-            model_name = meta["display_name"]
+        if meta.get("model_name"):
+            model_name = meta["model_name"]
 
-        extra_bits = []
-        if "publisher" in meta:
-            extra_bits.append(meta["publisher"])
-
-        if "params_string" in meta:
-            extra_bits.append(meta["params_string"])
-
-
-        model_name += f" {', '.join(extra_bits)}"
+        if meta["litellm_params"].get("model"):
+            model_name = "/".join(meta["litellm_params"].get("model").split("/")[1:])
 
         return model_name
